requirements.txt
black==22.10.0
boto3==1.26.8
botocore==1.29.8
click==8.1.3
jmespath==1.0.1
mypy-extensions==0.4.3
pathspec==0.10.2
platformdirs==2.5.4
python-dateutil==2.8.2
PyYAML==6.0
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.12
